move out expensive tests
